+++ app/src/main/java/com/example/mynotes/domain/usecases/FetchNote.kt
@@ -5,5 +5,5 @@ import javax.inject.Inject
 
 class FetchNote @Inject constructor(private val noteRepository: NoteRepository) {
 
-    suspend operator fun invoke(idNote: Long) = noteRepository.fetchNote(idNote = idNote)
+    suspend operator fun invoke(idNote: Int) = noteRepository.fetchNote(idNote = idNote)
 }
